+++ example/lib/native_app.dart
@@ -271,7 +271,8 @@ class _DemoEntry extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final style = SBBControlStyles.of(context);
+    final isLight = Theme.of(context).brightness == Brightness.light;
+    final color = SBBBaseStyle.resolve(isLight, SBBColors.white, SBBColors.charcoal);
     return OpenContainer(
       closedElevation: 0.0,
       openElevation: 0.0,
@@ -281,7 +282,7 @@ class _DemoEntry extends StatelessWidget {
       openShape: const RoundedRectangleBorder(),
       closedBuilder: (context, action) {
         return Container(
-          color: style.groupBackgroundColor,
+          color: color,
           child: SBBListItem(
             title: title,
             trailingIcon: SBBIcons.chevron_small_right_small,